Backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.5.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
